requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
flask==2.3.3
flask-cors==4.0.0
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
schedule==1.2.0
elasticsearch==8.11.0
bcrypt==4.0.1
scikit-learn==1.3.0
xgboost==2.0.0
tensorflow==2.13.0
joblib==1.3.1